requirements.txt
Django>=5.2,<6.0
djangorestframework
django-filter
django-cors-headers
pymysql
cryptography
dj-database-url
gunicorn
whitenoise
